refactor(account): extract Account Center auth redirect into useAuthRedirect
Move the two auth-redirect useEffect blocks out of App.tsx's Main
component into a dedicated useAuthRedirect hook. App.tsx now passes only
the URL-derived flags (isInCallback, isSilentAuthFailed); the hook owns
the SDK and PageContext reads, the redirect URI, and the silent re-auth
two-phase flow.

Side effects:
- App.tsx loses ~50 lines, bringing it back under the max-lines limit
  without needing a file-level eslint-disable.
- App.test.tsx keeps working unchanged because it renders <Main /> and
  the hook composes naturally; the existing mocks for @logto/react cover
  both call sites.
